Refactor some code out of PubmedArticleDeposit for it to be reused by other activities.
